Consolidated Supervisor API tests into clearer files
Change-type: patch
